Pass launch context to alignment api
closes OUT-6212 OUT6213
flag=none

I had to move the loading of the launch contexts to the alignment
components since that is now needed when fetching alignment set.
It is needed because we use it to fetch metadata about the
alignments.

This commit also renames sharedContexts to launchContexts. This was
done because of the change in direction regarding what contexts to
pull outcomes from.

You will also see test that have changed the expect.
 i.e expect(something) changed to expect(something).to.be.true

If you do not chain method calls after the expect, the test case
will pass regardless if what is in the expect statement.

Test Plan:
- requires g/341533
- using the demo app, inspect the page when loading the page.
- Go to the network tab. Find all request to api/alignment_sets/
- For Quiz XX, verify that launch context in in the payload of the
  request
  - Verify that includes[] includes the source_context_info
  - Verify that the Responce includes source_context_info,
    in_launch_context, launch_context_type
- Change the set of outcomes aligned with Quiz XX and select
  confirm alignments. Validate that the request to create
  alignment_set also has the launch context and source_context_info
- Validating the unit test are sufficient
- Exploratory testing
 - Try to break the demo app. This was a big change to outcomes-ui
   so make sure it works :)

The source context info and launch context data will be used when
implementing OUT-6085

Change-Id: Id79175e69ccc86dc8692e713d63fec1329a66346
Reviewed-on: https://gerrit.instructure.com/c/outcomes-ui/+/341764
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Martin Yosifov <martin.yosifov@instructure.com>
QA-Review: Chrystal Langston <chrystal.langston@instructure.com>
Product-Review: Kyle Rosenbaum <krosenbaum@instructure.com>